+++ src/game/objects/enemy/IEnemy.java
@@ -3,5 +3,5 @@ package game.objects.enemy;
 public interface IEnemy {
 
     public int getDamage();
-    public void hit(int damageAmount);
+    public void hit(double damageAmount);
 }
